docboxrx-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.128.0"}
psycopg = {extras = ["binary", "pool"], version = "^3.3.2"}
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.21"
pydantic-settings = "^2.12.0"
bcrypt = "^5.0.0"
cerebras-cloud-sdk = "^1.64.1"
sqlalchemy = "^2.0.45"
asyncpg = "^0.31.0"
aiosqlite = "^0.22.1"
nylas = "^6.14.0"
python-dotenv = "^1.2.1"
email-validator = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
